+++ tests/MessageTest.php
@@ -686,6 +686,22 @@ final class MessageTest extends AbstractTest
         $this->assertNull($inline->getFilename());
     }
 
+    public function testMissingFromHeader()
+    {
+        $this->mailbox->addMessage($this->getFixture('missing_from'));
+        $message = $this->mailbox->getMessage(1);
+
+        $this->assertNull($message->getFrom());
+    }
+
+    public function testMissingDateHeader()
+    {
+        $this->mailbox->addMessage($this->getFixture('missing_date'));
+        $message = $this->mailbox->getMessage(1);
+
+        $this->assertNull($message->getDate());
+    }
+
     public function testAttachmentMustNotBeCharsetDecoded()
     {
         $parts = [];